Refactor start script template to better support alternate implementations. Need to consider using a real template engine to reduce maintenance headaches keeping bash and bat versions in sync.
